+++ mulog-core/test/com/brunobonacci/mulog/levels_test.clj
@@ -5,22 +5,22 @@
 (fact
   "->filter does what it says"
 
-  ((lvl/->filter nil)           [{:id 1 :mulog/level ::lvl/info}])
+  (sequence (lvl/->filter nil)           [{:id 1 :mulog/level ::lvl/info}])
   => [{:id 1 :mulog/level ::lvl/info}]
 
-  ((lvl/->filter ::lvl/warning) [{:id 1 :mulog/level ::lvl/info}])
+  (sequence (lvl/->filter ::lvl/warning) [{:id 1 :mulog/level ::lvl/info}])
   => ()
 
-  ((lvl/->filter ::lvl/info)    [{:id 1 :mulog/level ::lvl/info}])
+  (sequence (lvl/->filter ::lvl/info)    [{:id 1 :mulog/level ::lvl/info}])
   => [{:id 1 :mulog/level ::lvl/info}]
 
-  ((lvl/->filter ::lvl/debug)   [{:id 1 :mulog/level ::lvl/info}])
+  (sequence (lvl/->filter ::lvl/debug)   [{:id 1 :mulog/level ::lvl/info}])
   => [{:id 1 :mulog/level ::lvl/info}]
 
   (let [ad-hoc (-> (make-hierarchy)
                    (derive ::bar ::foo)
                    (derive ::baz ::bar))]
-    ((lvl/->filter ::bar ad-hoc :some)
+    (sequence (lvl/->filter ::bar ad-hoc :some)
               [{:id 1 :some ::foo}
                {:id 2 :some ::bar}
                {:id 3 :some ::bar}