+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.1.9 - (unreleased)
   - Added flag to support ElasticSearch v6.x correctly
   - Added built-in leveled logging
+  - Deprecated `:transform` in favor of transducers via `:transduce`
 
 
 ## v0.1.8 - (2020-03-09)